+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.11.1 (unreleased)
 
-- Add `--repeat-result`
+- Add `--repeat-result` option to return the best, the worst or an average result with `--repeat-count`
+- Add `BenchmarkDriver::BulkOutput` to make an output plugin casually
 
 # v0.11.0